Point to github.com/billy99/submariner for updated Endpoint CRD
The Submariner Endpoint CRD was updated to include the AllocatedIPs from the
ClusterGlobalEgressIPs. The Endpoint CRD struct is define in the Submariner
repo, so need to point to a version that contains the change.

This alos pulled in the latest Shipyard.

Temporarily point to astoycos version. Once his code merges, point to feature
branch.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>